backend/requirements.txt
# VoicePrep AI Backend Dependencies

# Web Framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0

# Pydantic for data validation
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.4.0,<3.0.0

# HTTP Client
httpx>=0.27.0,<1.0.0
aiohttp>=3.10.0,<4.0.0

# Google Cloud
google-cloud-aiplatform>=1.60.0,<2.0.0
google-cloud-storage>=3.0.0  # Upgraded to fix deprecation warning
google-cloud-speech>=2.25.0,<3.0.0  # Speech-to-Text API for voice input
google-auth>=2.30.0,<3.0.0
google-genai>=1.0.0  # Unified SDK for Vertex AI and API key (replaces google-generativeai)

# Firebase
firebase-admin>=6.5.0,<7.0.0

# Encryption & Security
google-cloud-kms>=2.21.0,<3.0.0
google-cloud-secret-manager>=2.20.0,<3.0.0
cryptography>=42.0.0,<43.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Document Processing
python-multipart>=0.0.9,<1.0.0
pypdf>=4.0.0,<7.0.0
python-docx>=1.1.0,<2.0.0

# PDF Generation (for improved resume export)
reportlab>=4.0.0,<5.0.0
markdown>=3.5.0,<4.0.0

# Utilities
python-dotenv>=1.0.0,<2.0.0

# Development (optional - install with: pip install -r requirements-dev.txt)
# pytest>=8.0.0
# pytest-asyncio>=0.24.0
# ruff>=0.6.0
